JobFit_Multi-Agent/backend/requirements.txt
# JobFit_Multi-Agent 后端依赖

# Web框架
Flask==2.3.3
flask-cors==4.0.0

# LLM API调用
openai==1.3.5
requests==2.31.0

# 文档解析
PyPDF2==3.0.1
python-docx==1.0.1
chardet==5.2.0

# OCR识别
pytesseract==0.3.10
Pillow==10.0.1

# 环境变量管理
python-dotenv==1.0.0

# 数据处理
json5==0.9.14
